Add Repositories, IoC

+++ SerwisProduktow.Domain/Entities/Rating.cs
@@ -17,7 +17,7 @@ namespace SerwisProduktow.Domain.Entities
         {
             SumOfVotes =SumOfVotes + ocena;
             NumberOfVotes++;
-            AvarageOfVotes = SumOfVotes / NumberOfVotes;
+            AvarageOfVotes = (double)SumOfVotes / (double)NumberOfVotes;
         }
     }
 }
